requirements.txt
requests==2.31.0
bs4==0.0.2
tweepy==4.14.0
google-generativeai==0.3.2